automation/social_media/requirements.txt
requests>=2.31.0
tweepy>=4.14.0
python-dotenv>=1.0.0
flask>=3.0.0
gunicorn>=21.2.0
werkzeug>=3.0.1
